Bedarf Domäne refactored

+++ src/main/java/io/remedymatch/bedarf/api/BedarfDTO.java
@@ -1,12 +1,14 @@
 package io.remedymatch.bedarf.api;
 
+import java.math.BigDecimal;
+import java.util.UUID;
+
+import javax.validation.constraints.NotNull;
+
 import io.remedymatch.artikel.api.ArtikelDTO;
 import io.remedymatch.institution.api.InstitutionStandortDTO;
 import lombok.Builder;
 import lombok.Data;
-
-import javax.validation.constraints.NotNull;
-import java.util.UUID;
 
 @Data
 @Builder
@@ -18,7 +20,7 @@ public class BedarfDTO {
     private ArtikelDTO artikel;
 
     @NotNull
-    private double anzahl;
+    private BigDecimal anzahl;
 
     @NotNull
     private InstitutionStandortDTO standort;
@@ -28,7 +30,7 @@ public class BedarfDTO {
 
     private UUID institutionId;
 
-    private double rest;
+    private BigDecimal rest;
 
     private boolean steril;
 
@@ -38,6 +40,6 @@ public class BedarfDTO {
 
     private boolean bedient;
 
-    private double entfernung;
+    private BigDecimal entfernung;
 
 }
